requirements.txt
jupyter==1.0.0
matplotlib==3.7.1
networkx==2.8.4
numpy==1.23.5
pyg==2.0.4
python==3.9.12
pytorch==1.11.0
seaborn==0.13.2
tqdm==4.64.1
